+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/events/salt/NodeNotifierImpl.java
@@ -68,4 +68,9 @@ public class NodeNotifierImpl extends NodeImpl implements Node, NotifyingLabelab
   public void removeAll() {
     applyRemoveAllLabels(super::removeAll, this);
   }
+
+  @Override
+  public void setId(String id) {
+    setNotficiationAwareId(this, id);
+  }
 }